+++ src/main/java/dev/slimevr/filtering/CircularArrayList.java
@@ -1,4 +1,4 @@
-package dev.slimevr.math;
+package dev.slimevr.filtering;
 
 import java.util.*;
