+++ src/main/java/dev/slimevr/vr/trackers/IMUTracker.java
@@ -5,7 +5,8 @@ import com.jme3.math.Quaternion;
 import com.jme3.math.Vector3f;
 import dev.slimevr.VRServer;
 import dev.slimevr.config.TrackerConfig;
-import dev.slimevr.math.QuaternionMovingAverage;
+import dev.slimevr.filtering.QuaternionMovingAverage;
+import dev.slimevr.filtering.TrackerFilters;
 import dev.slimevr.vr.Device;
 import dev.slimevr.vr.trackers.udp.TrackersUDPServer;
 import dev.slimevr.vr.trackers.udp.UDPDevice;
