+++ src/main/java/dev/slimevr/vr/trackers/TrackerWithFiltering.java
@@ -1,6 +1,8 @@
 package dev.slimevr.vr.trackers;
 
 
+import dev.slimevr.filtering.TrackerFilters;
+
 public interface TrackerWithFiltering {
 
 	void setFiltering(TrackerFilters type, float amount, int buffer);
